+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the create-aptos-dapp tool will be captured in this file.
 
 # Unreleased
 
+- Install Aptos CLI after dependencies are installed
+
 # 0.0.31 (2024-09-30)
 
 - Add new custom indexer template